requirements.txt
protobuf==3.17.3
fastapi[all]==0.66.1
sentencepiece==0.1.96
transformers[torch]==4.7.0
Pillow==8.0